requirements.txt
requests==2.13.0
pyyaml==3.12
pyopenssl==17.0.0
lxml==3.7.3
ndg-httpsclient==0.4.2
pyasn1==0.2.3
beautifulsoup4==4.5.3
pillow==2.0.0